auth-service/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
Werkzeug==2.0.2
PyJWT==2.3.0
